+++ test/binary_counter_test.dart
@@ -1,5 +1,6 @@
 import 'package:binary_counter/binary_counter.dart';
 import 'package:test/test.dart';
+import 'package:number_system/number_system.dart';
 
 void main() {
   test("Random binary number", (){
@@ -8,15 +9,23 @@ void main() {
     ));
   });
   test("Binary addition", (){
-    print(binaryAddition(
-      binaryNumber1: "0101", 
-      binaryNumber2: "1111",
-    ));
+    String binaryNumber = "0";
+    for(int i = 0; i < 10; i++){
+      binaryNumber = binaryAddition(
+        binaryNumber1: binaryNumber, 
+        binaryNumber2: "1",
+      );
+      print(binaryNumber.binaryToDec());
+    }
   });
   test("Binary subtraction", (){
     print(binarySubtraction(
-      binaryNumber: "11001010", 
-      binaryAmountToSubtract: "10011011",
+      binaryNumber: "10", 
+      binaryAmountToSubtract: "1",
     ));
   });
+  test("Binary multiplication", ((){
+    String result = binaryMultiplication(binaryNumber: "10", times: "10");
+    print(result);
+  }));
 }